fix a bunch of remaining warnings
git-svn-id: https://svn.code.sf.net/p/pinmame/code/trunk@3928 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf

+++ src/cheat.c
@@ -10082,7 +10082,7 @@ static void DoCheatAction(CheatAction * action)
 
 				if(currentValue != action->lastValue)
 				{
-					action->frameTimer = parameter * Machine->drv->frames_per_second;
+					action->frameTimer = (INT32)(parameter * Machine->drv->frames_per_second);
 
 					action->flags |= kActionFlag_WasModified;
 				}
